[#75] refactor: auth 파일 이름와 디렉토리 변경
auth 파일 action에서 utils로 이동했습니다.
